requirements.txt
Django==4.2.13
django-froala-editor
